Compose full report document context
