+++ apps/server/tests/adapters/pdf/test_report_mapping_pipeline.py
@@ -51,11 +51,7 @@ def _prepared_report_input() -> PreparedReportInput:
 
 def test_prepare_report_input_returns_mapping_ready_boundary_types() -> None:
     prepared = _prepared_report_input()
-    composition = report_document.compose_report_document(
-        aggregate=prepared.domain_test_run,
-        report_facts=prepared.report_facts,
-        lang=prepared.language,
-    )
+    context = report_document.compose_report_document_context(prepared)
 
     assert isinstance(prepared, shared_reporting.PreparedReportInput)
     assert isinstance(prepared.report_facts, shared_reporting.PreparedReportFacts)
@@ -63,7 +59,7 @@ def test_prepare_report_input_returns_mapping_ready_boundary_types() -> None:
         prepared.report_facts.decision.primary_candidate,
         shared_report_projection.PrimaryReportFacts,
     )
-    assert isinstance(composition, report_document.ReportDocumentComposition)
+    assert isinstance(context, shared_report_document.ReportDocumentContext)
     assert not hasattr(prepared, "renderer_payload")
     assert not hasattr(prepared, "presentation")
