+++ apps/server/vibesensor/use_cases/history/report_document/sections.py
@@ -17,7 +17,6 @@ from vibesensor.shared.boundaries.reporting.document import (
 )
 from vibesensor.use_cases.history.report_document._candidate_resolver import PrimaryCandidateContext
 
-from .composition import ReportDocumentComposition
 from .measurements import _evidence_chain_rows
 from .narrative_summaries import (
     _context_summary_text,
@@ -94,7 +93,8 @@ def _build_appendix_c_data(
     aggregate: TestRun,
     measurements: list[MeasurementRow],
     report_facts: PreparedReportFacts,
-    composition: ReportDocumentComposition,
+    speed_window_label: str | None,
+    proof_caveat: str | None,
     data_trust: list[DataTrustItem],
     tr: Callable[..., str],
 ) -> AppendixCData:
@@ -113,7 +113,8 @@ def _build_appendix_c_data(
         context_summary=_context_summary_text(primary, report_facts, tr=tr),
         limits_summary=_run_limits_summary_text(
             report_facts,
-            composition,
+            speed_window_label=speed_window_label,
+            proof_caveat=proof_caveat,
             tr=tr,
         ),
         speed_band_summary=speed_summary,
